Updated elaxer/chess@v1.4.1

+++ internal/piece/bishop.go
@@ -34,7 +34,7 @@ func (b *Bishop) Notation() string {
 	return NotationBishop
 }
 
-func (b *Bishop) Weight() uint8 {
+func (b *Bishop) Weight() uint16 {
 	return WeightBishop
 }
